# GYMaster-API/DTO/AsistenciasDTO.cs
namespace GYMaster_API.DTO
{
    public class AsistenciasDTO
    {
        public string UsuarioUsuarios { get; set; } = null!;

        public DateTime Fecha { get; set; }

        public TimeSpan Hora { get; set; }
    }
}