# GYMaster-API/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "GYMaster",
    "version": "v1"
  },
  "paths": {
    "/api/Asistencias/Buscar": {
      "get": {
        "tags": [
          "Asistencias"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "UsuarioUsuarios",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/AsistenciasDTO"
              }
            }
          }
        }
      }
    },
    "/api/Asistencias/Catalogo": {
      "get": {
        "tags": [
          "Asistencias"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "UsuarioGimnasios",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/AsistenciasDTO"
              }
            }
          }
        }
      }
    },
    "/api/Asistencias/Registrar": {
      "post": {
        "tags": [
          "Asistencias"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/AsistenciasDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Carritosdetalles/Catalogo": {
      "get": {
        "tags": [
          "Carritosdetalles"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "UsuarioUsuarios",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/CarritosdetalleDTO"
              }
            }
          }
        }
      }
    },
    "/api/Carritosdetalles/Agregar": {
      "post": {
        "tags": [
          "Carritosdetalles"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CarritosdetalleDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Carritosdetalles/Ajustar": {
      "put": {
        "tags": [
          "Carritosdetalles"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CarritosdetalleDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Carritosmaestros/Buscar": {
      "get": {
        "tags": [
          "Carritosmaestros"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "UsuarioUsuarios",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/CarritosmaestroDTO"
            }
          }
        }
      }
    },
    "/api/Carritosmaestros/Registrar": {
      "post": {
        "tags": [
          "Carritosmaestros"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "UsuarioUsuarios",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Carritosmaestros/Actualizar": {
      "patch": {
        "tags": [
          "Carritosmaestros"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CarritosmaestroDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
    "AsistenciasDTO": {
      "type": "object",
      "properties": {
        "usuarioUsuarios": {
          "type": "string"
        },
        "fecha": {
          "format": "date-time",
          "type": "string"
        },
        "hora": {
          "$ref": "#/definitions/TimeSpan"
        }
      }
    },
    "CarritosdetalleDTO": {
      "type": "object",
      "properties": {
        "usuarioUsuarios": {
          "type": "string"
        },
        "idProducto": {
          "format": "int32",
          "type": "integer"
        },
        "cantidad": {
          "format": "int32",
          "type": "integer"
        },
        "subtotal": {
          "format": "double",
          "type": "number"
        }
      }
    },
    "CarritosmaestroDTO": {
      "type": "object",
      "properties": {
        "usuarioUsuarios": {
          "type": "string"
        },
        "total": {
          "format": "double",
          "type": "number"
        }
      }
    },
...
    "TimeSpan": {
      "type": "object",
      "properties": {
        "ticks": {
          "format": "int64",
          "type": "integer"
        },
        "days": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "hours": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "milliseconds": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "minutes": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "seconds": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "totalDays": {
          "format": "double",
          "type": "number",
          "readOnly": true
        },
        "totalHours": {
          "format": "double",
          "type": "number",
          "readOnly": true
        },
        "totalMilliseconds": {
          "format": "double",
          "type": "number",
          "readOnly": true
        },
        "totalMinutes": {
          "format": "double",
          "type": "number",
          "readOnly": true
        },
        "totalSeconds": {
          "format": "double",
          "type": "number",
          "readOnly": true
        }
      }
    },
...
  }
}